Arrange into modules / extensions and get test passing
