+++ lib/active_record/filter/unkown_filter_error.rb
@@ -1,0 +1,2 @@
+class ActiveRecord::UnkownFilterError < NoMethodError
+end
